blog_page_commit2
