+++ blog_page/src/app/components/cards.jsx
@@ -1,36 +1,109 @@
 import './cards_style.css'
-export default function Card()
+export default function DaisyCards()
 {
     return(
-    <>
-    <div className='blogs'>
-    <div className="card-container">
-        <div className='image'></div>
-        <div className='content'></div>
-    </div>
-    <div className="card-container">
-        <div className='image'></div>
-        <div className='content'></div>
-    </div>
-    <div className="card-container">
-        <div className='image'></div>
-        <div className='content'></div>
-    </div>
-    </div>
-    <div className='blogs'>
-    <div className="card-container">
-        <div className='image'></div>
-        <div className='content'></div>
-    </div>
-    <div className="card-container">
-        <div className='image'></div>
-        <div className='content'></div>
-    </div>
-    <div className="card-container">
-        <div className='image'></div>
-        <div className='content'></div>
-    </div>
-    </div>
-    </>
+        <>
+            <div className="card-container">
+                <div className="card w-90 bg-base-100 shadow-xl mr-20px">
+                    <figure><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTRvzY6Y9QLcN_il-84ikH7W8iZpXX6rHY2tjMhRLhgoP_YK_fIaRl7QLnnOg&s" alt="Shoes" /></figure>
+                    <div className="card-body">
+                        <h2 className="card-title">
+                        <span className='gradient'>#Blog1</span>
+                        <div className="badge badge-secondary">NEW</div>
+                        </h2>
+                        <p>Brief Description of the blog</p>
+                        <div className="card-actions justify-start">
+                        <div className="badge b1 badge-outline">Tech</div> 
+                        <div className="badge b2 badge-outline">Marketing</div>
+                        <div className="badge b3 badge-outline">Start-up</div>
+                        </div>
+                        <button className="btn">Read More<svg class="w-4 h-4 text-white-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 10 16"><path d="M3.414 1A2 2 0 0 0 0 2.414v11.172A2 2 0 0 0 3.414 15L9 9.414a2 2 0 0 0 0-2.828L3.414 1Z"/></svg></button>
+                    </div>
+                </div>
+                <div className="card w-90 bg-base-100 shadow-xl">
+                    <figure><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTRvzY6Y9QLcN_il-84ikH7W8iZpXX6rHY2tjMhRLhgoP_YK_fIaRl7QLnnOg&s" alt="Shoes" /></figure>
+                    <div className="card-body">
+                        <h2 className="card-title">
+                        <span className='gradient'>#Blog2</span>
+                        <div className="badge badge-secondary">NEW</div>
+                        </h2>
+                        <p>Brief Description of the blog</p>
+                        <div className="card-actions justify-start">
+                        <div className="badge b1 badge-outline">Tech</div> 
+                        <div className="badge b2 badge-outline">Marketing</div>
+                        <div className="badge b3 badge-outline">Start-up</div>
+                        </div>
+                        <button className="btn">Read More<svg class="w-4 h-4 text-white-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 10 16"><path d="M3.414 1A2 2 0 0 0 0 2.414v11.172A2 2 0 0 0 3.414 15L9 9.414a2 2 0 0 0 0-2.828L3.414 1Z"/></svg></button>
+                    </div>
+                </div>
+                <div className="card w-90 bg-base-100 shadow-xl">
+                    <figure><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTRvzY6Y9QLcN_il-84ikH7W8iZpXX6rHY2tjMhRLhgoP_YK_fIaRl7QLnnOg&s" alt="Shoes" /></figure>
+                    <div className="card-body">
+                        <h2 className="card-title">
+                        <span className='gradient'>#Blog3</span>
+                        <div className="badge badge-secondary">NEW</div>
+                        </h2>
+                        <p>Brief Description of the blog</p>
+                        <div className="card-actions justify-start">
+                        <div className="badge b1 badge-outline">Tech</div> 
+                        <div className="badge b2 badge-outline">Marketing</div>
+                        <div className="badge b3 badge-outline">Start-up</div>
+                        </div>
+                        <button className="btn">Read More<svg class="w-4 h-4 text-white-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 10 16"><path d="M3.414 1A2 2 0 0 0 0 2.414v11.172A2 2 0 0 0 3.414 15L9 9.414a2 2 0 0 0 0-2.828L3.414 1Z"/></svg></button>
+                    </div>
+                </div>
+            </div>
+            <div className="card-container">
+                <div className="card w-90 bg-base-100 shadow-xl">
+                    <figure><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTRvzY6Y9QLcN_il-84ikH7W8iZpXX6rHY2tjMhRLhgoP_YK_fIaRl7QLnnOg&s" alt="Shoes" /></figure>
+                    <div className="card-body">
+                        <h2 className="card-title">
+                        <span className='gradient'>#Blog4</span>
+                        <div className="badge badge-secondary">NEW</div>
+                        </h2>
+                        <p>Brief Description of the blog</p>
+                        <div className="card-actions justify-start">
+                        <div className="badge b1 badge-outline">Tech</div> 
+                        <div className="badge b2 badge-outline">Marketing</div>
+                        <div className="badge b3 badge-outline">Start-up</div>
+                        </div>
+                        <button className="btn">Read More<svg class="w-4 h-4 text-white-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 10 16"><path d="M3.414 1A2 2 0 0 0 0 2.414v11.172A2 2 0 0 0 3.414 15L9 9.414a2 2 0 0 0 0-2.828L3.414 1Z"/></svg></button>
+                    </div>
+                </div>
+                <div className="card w-90 bg-base-100 shadow-xl">
+                    <figure><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTRvzY6Y9QLcN_il-84ikH7W8iZpXX6rHY2tjMhRLhgoP_YK_fIaRl7QLnnOg&s" alt="Shoes" /></figure>
+                    <div className="card-body">
+                        <h2 className="card-title">
+                        <span className='gradient'>#Blog5</span>
+                        <div className="badge badge-secondary">NEW</div>
+                        </h2>
+                        <p>Brief Description of the blog</p>
+                        <div className="card-actions justify-start">
+                        <div className="badge b1 badge-outline">Tech</div> 
+                        <div className="badge b2 badge-outline">Marketing</div>
+                        <div className="badge b3 badge-outline">Start-up</div>
+                        </div>
+                        <button className="btn">Read More<svg class="w-4 h-4 text-white-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 10 16"><path d="M3.414 1A2 2 0 0 0 0 2.414v11.172A2 2 0 0 0 3.414 15L9 9.414a2 2 0 0 0 0-2.828L3.414 1Z"/></svg></button>
+                    </div>
+                </div>
+                <div className="card w-90 bg-base-100 shadow-xl">
+                    <figure><img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTRvzY6Y9QLcN_il-84ikH7W8iZpXX6rHY2tjMhRLhgoP_YK_fIaRl7QLnnOg&s" alt="Shoes" /></figure>
+                    <div className="card-body">
+                        <h2 className="card-title">
+                        <span className='gradient'>#Blog6</span>
+                        <div className="badge badge-secondary">NEW</div>
+                        </h2>
+                        <p>Brief Description of the blog</p>
+                        <div className="card-actions justify-start">
+                        <div className="badge b1 badge-outline">Tech</div> 
+                        <div className="badge b2 badge-outline">Marketing</div>
+                        <div className="badge b3 badge-outline">Start-up</div>
+                        </div>
+                        <button className="btn">Read More<svg class="w-4 h-4 text-white-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 10 16"><path d="M3.414 1A2 2 0 0 0 0 2.414v11.172A2 2 0 0 0 3.414 15L9 9.414a2 2 0 0 0 0-2.828L3.414 1Z"/></svg>
+                        </button>
+                    </div>
+                </div>
+            </div>
+        </>
     )
 }
