--- blog_page/src/app/components/cards1.jsx
@@ -1,108 +1,0 @@
-import './cards_style.css'
-export default function DaisyCards()
-{
-    return(
-        <>
-            <div className="card-container">
-                <div className="card w-90 bg-base-100 shadow-xl mr-20px">
-                    <figure><img src="https://daisyui.com/images/stock/photo-1606107557195-0e29a4b5b4aa.jpg" alt="Shoes" /></figure>
-                    <div className="card-body">
-                        <h2 className="card-title">
-                        #Blog1
-                        <div className="badge badge-secondary">NEW</div>
-                        </h2>
-                        <p>Brief Description of the blog</p>
-                        <div className="card-actions justify-start">
-                        <div className="badge b1 badge-outline">Fashion</div> 
-                        <div className="badge b2 badge-outline">Products</div>
-                        <div className="badge b3 badge-outline">Products</div>
-                        </div>
-                        <button className="btn">Read More</button>
-                    </div>
-                </div>
-                <div className="card w-90 bg-base-100 shadow-xl">
-                    <figure><img src="https://daisyui.com/images/stock/photo-1606107557195-0e29a4b5b4aa.jpg" alt="Shoes" /></figure>
-                    <div className="card-body">
-                        <h2 className="card-title">
-                        #Blog2
-                        <div className="badge badge-secondary">NEW</div>
-                        </h2>
-                        <p>Brief Description of the blog</p>
-                        <div className="card-actions justify-start">
-                        <div className="badge b1 badge-outline">Fashion</div> 
-                        <div className="badge b2 badge-outline">Products</div>
-                        <div className="badge b3 badge-outline">Products</div>
-                        </div>
-                        <button className="btn">Read More</button>
-                    </div>
-                </div>
-                <div className="card w-90 bg-base-100 shadow-xl">
-                    <figure><img src="https://daisyui.com/images/stock/photo-1606107557195-0e29a4b5b4aa.jpg" alt="Shoes" /></figure>
-                    <div className="card-body">
-                        <h2 className="card-title">
-                        #Blog3
-                        <div className="badge badge-secondary">NEW</div>
-                        </h2>
-                        <p>Brief Description of the blog</p>
-                        <div className="card-actions justify-start">
-                        <div className="badge b1 badge-outline">Fashion</div> 
-                        <div className="badge b2 badge-outline">Products</div>
-                        <div className="badge b3 badge-outline">Products</div>
-                        </div>
-                        <button className="btn">Read More</button>
-                    </div>
-                </div>
-            </div>
-            <div className="card-container">
-                <div className="card w-90 bg-base-100 shadow-xl">
-                    <figure><img src="https://daisyui.com/images/stock/photo-1606107557195-0e29a4b5b4aa.jpg" alt="Shoes" /></figure>
-                    <div className="card-body">
-                        <h2 className="card-title">
-                        #Blog4
-                        <div className="badge badge-secondary">NEW</div>
-                        </h2>
-                        <p>Brief Description of the blog</p>
-                        <div className="card-actions justify-start">
-                        <div className="badge b1 badge-outline">Fashion</div> 
-                        <div className="badge b2 badge-outline">Products</div>
-                        <div className="badge b3 badge-outline">Products</div>
-                        </div>
-                        <button className="btn">Read More</button>
-                    </div>
-                </div>
-                <div className="card w-90 bg-base-100 shadow-xl">
-                    <figure><img src="https://daisyui.com/images/stock/photo-1606107557195-0e29a4b5b4aa.jpg" alt="Shoes" /></figure>
-                    <div className="card-body">
-                        <h2 className="card-title">
-                        #Blog5
-                        <div className="badge badge-secondary">NEW</div>
-                        </h2>
-                        <p>Brief Description of the blog</p>
-                        <div className="card-actions justify-start">
-                        <div className="badge b1 badge-outline">Fashion</div> 
-                        <div className="badge b2 badge-outline">Products</div>
-                        <div className="badge b3 badge-outline">Products</div>
-                        </div>
-                        <button className="btn">Read More</button>
-                    </div>
-                </div>
-                <div className="card w-90 bg-base-100 shadow-xl">
-                    <figure><img src="https://daisyui.com/images/stock/photo-1606107557195-0e29a4b5b4aa.jpg" alt="Shoes" /></figure>
-                    <div className="card-body">
-                        <h2 className="card-title">
-                        #Blog6
-                        <div className="badge badge-secondary">NEW</div>
-                        </h2>
-                        <p>Brief Description of the blog</p>
-                        <div className="card-actions justify-start">
-                        <div className="badge b1 badge-outline">Fashion</div> 
-                        <div className="badge b2 badge-outline">Products</div>
-                        <div className="badge b3 badge-outline">Products</div>
-                        </div>
-                        <button className="btn">Read More</button>
-                    </div>
-                </div>
-            </div>
-        </>
-    )
-}
