+++ blog_page/src/app/components/cards_style.css
@@ -3,6 +3,10 @@
     flex-direction: row;
     margin:20px;
 }
+img{
+    width:500px;
+    height:300px
+}
 .card{
     margin:15px;
     background-color: black;
@@ -13,27 +17,28 @@
 .b1{
     background-color: rgb(111, 0, 255);
     border-color:rgb(111, 0, 255) ;
+    height:30px;
 }
 .b2{
     background-color: rgb(0, 153, 255);
     border-color:rgb(0, 153, 255) ;
+    height:30px;
 }
 .b3{
     background-color: rgb(24, 13, 167);
     border-color:rgb(24, 13, 167);
+    height:30px;
 }
-h1{
-    background: -webkit-linear-gradient(45deg, #ff6600, #ff00ff); /* For Safari 5.1 to 6.0 */
-      background: linear-gradient(45deg, #ff6600, #ff00ff); /* Standard syntax */
-      -webkit-background-clip: text; /* Clip the background to the text */
-      -webkit-text-fill-color: transparent; /* Hide the original text color */
-}
+
 .card:hover{
     box-shadow: -5px -5px 5px rgba(37, 37, 37, 0.979);
-    transition: 0.2s ease-out;
+    transition: 0.3s ease-out;
 }
 .card-actions{
     margin-bottom:20px;
+}
+.btn{
+    font-size: 15px;
 }
 .btn:hover{
     background-color: rgb(41, 41, 77);
