+++ blog_page/src/app/globals.css
@@ -9,6 +9,17 @@ body {
   background: black;
 }
 
+.gradient{
+  background-color: rgb(0, 208, 255);
+  background-image: linear-gradient(30deg,#78e9f3, #1c4283);
+  background-size: 100%;
+  background-repeat: repeat;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent; 
+  -moz-background-clip: text;
+  -moz-text-fill-color: transparent;
+}
+
 @layer utilities {
   .text-balance {
     text-wrap: balance;
